requirements.txt
fastapi
numpy
pandas
psycopg2-binary
pyarrow
scikit-learn
uvicorn